+++ src/com/intellij/idea/plugin/hybris/impex/lang/folding/util/ImpExSmartFoldingPlaceholderBuilder.kt
@@ -1,7 +1,7 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -64,19 +64,28 @@ class ImpExSmartFoldingPlaceholderBuilder : ImpexFoldingPlaceholderBuilder {
                 ?.text
                 ?: impexAttribute.text
             return "!$foldedText"
-        }
-        if (quoteAwareStringEquals(text,
-                AttributeModifier.LANG,
-                AttributeModifier.DATE_FORMAT,
-                AttributeModifier.MODE,
-                AttributeModifier.NUMBER_FORMAT)) {
+        } else if (quoteAwareStringEquals(text, TypeModifier.DISABLE_INTERCEPTOR_BEANS)) {
             return impexAttribute.anyAttributeValue
                 ?.text
                 ?: impexAttribute.text
-        } else if (quoteAwareStringEquals(text,
+        } else if (quoteAwareStringEquals(
+                text,
+                AttributeModifier.LANG,
+                AttributeModifier.DATE_FORMAT,
+                AttributeModifier.MODE,
+                AttributeModifier.NUMBER_FORMAT
+            )
+        ) {
+            return impexAttribute.anyAttributeValue
+                ?.text
+                ?: impexAttribute.text
+        } else if (quoteAwareStringEquals(
+                text,
                 AttributeModifier.TRANSLATOR,
                 AttributeModifier.CELL_DECORATOR,
-                TypeModifier.PROCESSOR)) {
+                TypeModifier.PROCESSOR
+            )
+        ) {
 
             val value = impexAttribute.anyAttributeValue?.text ?: return impexAttribute.text
             val clearedString = QUOTES_PATTERN.matcher(value).replaceAll(StringUtils.EMPTY)
@@ -84,7 +93,8 @@ class ImpExSmartFoldingPlaceholderBuilder : ImpexFoldingPlaceholderBuilder {
             return clearedString.substringAfterLast(".")
                 .ifBlank { value }
         } else {
-            return isBooleanAttributeModifier(impexAttribute,
+            return isBooleanAttributeModifier(
+                impexAttribute,
                 AttributeModifier.UNIQUE,
                 AttributeModifier.DEFAULT,
                 AttributeModifier.VIRTUAL,